+++ src/storm-dft/api/storm-dft.h
@@ -11,6 +11,7 @@
 #include "storm-dft/storage/DftJsonExporter.h"
 #include "storm-dft/transformations/DftToGspnTransformator.h"
 #include "storm-dft/transformations/DftTransformer.h"
+#include "storm-dft/utility/DftValidator.h"
 #include "storm-dft/utility/FDEPConflictFinder.h"
 #include "storm-dft/utility/FailureBoundFinder.h"
 #include "storm-dft/utility/RelevantEvents.h"
@@ -57,13 +58,18 @@ std::shared_ptr<storm::dft::storage::DFT<ValueType>> loadDFTJsonFile(std::string
  * Check whether the DFT is well-formed.
  *
  * @param dft DFT.
- * @param validForAnalysis  If true, additional (more restrictive) checks are performed to check whether the DFT is valid for analysis.
- * @return Pair where the first entry is true iff the DFT is well-formed. The second entry contains the error messages for illformed parts.
+ * @param validForMarkovianAnalysis If true, additional checks are performed to check whether the DFT is valid for analysis via Markov models.
+ * @return Pair where the first entry is true iff the DFT is well-formed. The second entry contains the error messages for ill-formed parts.
  */
 template<typename ValueType>
-std::pair<bool, std::string> isWellFormed(storm::dft::storage::DFT<ValueType> const& dft, bool validForAnalysis = true) {
+std::pair<bool, std::string> isWellFormed(storm::dft::storage::DFT<ValueType> const& dft, bool validForMarkovianAnalysis = true) {
     std::stringstream stream;
-    bool wellFormed = dft.checkWellFormedness(validForAnalysis, stream);
+    bool wellFormed = false;
+    if (validForMarkovianAnalysis) {
+        wellFormed = storm::dft::utility::DftValidator<ValueType>::isDftValidForMarkovianAnalysis(dft, stream);
+    } else {
+        wellFormed = storm::dft::utility::DftValidator<ValueType>::isDftWellFormed(dft, stream);
+    }
     return std::pair<bool, std::string>(wellFormed, stream.str());
 }
 
@@ -73,13 +79,14 @@ std::pair<bool, std::string> isWellFormed(storm::dft::storage::DFT<ValueType> co
  * @param dft DFT.
  * @param uniqueBE Flag whether a unique constant failed BE is created.
  * @param binaryFDEP Flag whether all dependencies should be binary (only one dependent child).
+ * @param exponentialDistributions Flag whether distributions should be transformed to exponential distributions (if possible).
  * @return Transformed DFT.
  */
 template<typename ValueType>
 std::shared_ptr<storm::dft::storage::DFT<ValueType>> applyTransformations(storm::dft::storage::DFT<ValueType> const& dft, bool uniqueBE, bool binaryFDEP,
-                                                                          bool markovianDistributions) {
+                                                                          bool exponentialDistributions) {
     std::shared_ptr<storm::dft::storage::DFT<ValueType>> transformedDft = std::make_shared<storm::dft::storage::DFT<ValueType>>(dft);
-    if (markovianDistributions && !storm::dft::transformations::DftTransformer<ValueType>::hasOnlyExponentialDistributions(*transformedDft)) {
+    if (exponentialDistributions && !storm::dft::transformations::DftTransformer<ValueType>::hasOnlyExponentialDistributions(*transformedDft)) {
         transformedDft = storm::dft::transformations::DftTransformer<ValueType>::transformExponentialDistributions(*transformedDft);
     }
     if (uniqueBE && !storm::dft::transformations::DftTransformer<ValueType>::hasUniqueFailedBE(*transformedDft)) {